Support replication of generated column during initial sync
When 'copy_data' is true, during the initial sync, the data is replicated from
the publisher to the subscriber using the COPY command. The normal COPY
command does not copy generated columns, so when 'include_generated_columns'
is true, we need to copy using the syntax:
'COPY (SELECT column_name FROM table_name) TO STDOUT'.

Note that we don't copy columns when the subscriber-side column is also
generated. Those will be filled as normal with the subscriber-side computed
or default data.